jadx-1.1.0 - PS v8.0.1 (24289)

+++ sources/org/telegram/tgnet/TLRPC$Document.java
@@ -16,7 +16,7 @@ public abstract class TLRPC$Document extends TLObject {
     public String mime_type;
     public int size;
     public ArrayList<TLRPC$PhotoSize> thumbs = new ArrayList<>();
-    public int user_id;
+    public long user_id;
     public int version;
     public ArrayList<TLRPC$VideoSize> video_thumbs = new ArrayList<>();
