+++ sources/org/telegram/tgnet/TLRPC$TL_account_acceptAuthorization.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public class TLRPC$TL_account_acceptAuthorization extends TLObject {
     public static int constructor = -NUM;
-    public int bot_id;
+    public long bot_id;
     public TLRPC$TL_secureCredentialsEncrypted credentials;
     public String public_key;
     public String scope;
@@ -16,7 +16,7 @@ public class TLRPC$TL_account_acceptAuthorization extends TLObject {
 
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
-        abstractSerializedData.writeInt32(this.bot_id);
+        abstractSerializedData.writeInt64(this.bot_id);
         abstractSerializedData.writeString(this.scope);
         abstractSerializedData.writeString(this.public_key);
         abstractSerializedData.writeInt32(NUM);
