+++ sources/org/telegram/tgnet/TLRPC$TL_account_getAuthorizationForm.java
@@ -2,7 +2,7 @@ package org.telegram.tgnet;
 
 public class TLRPC$TL_account_getAuthorizationForm extends TLObject {
     public static int constructor = -NUM;
-    public int bot_id;
+    public long bot_id;
     public String public_key;
     public String scope;
 
@@ -12,7 +12,7 @@ public class TLRPC$TL_account_getAuthorizationForm extends TLObject {
 
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
-        abstractSerializedData.writeInt32(this.bot_id);
+        abstractSerializedData.writeInt64(this.bot_id);
         abstractSerializedData.writeString(this.scope);
         abstractSerializedData.writeString(this.public_key);
     }
