+++ sources/org/telegram/tgnet/TLRPC$TL_account_getThemes.java
@@ -3,7 +3,7 @@ package org.telegram.tgnet;
 public class TLRPC$TL_account_getThemes extends TLObject {
     public static int constructor = NUM;
     public String format;
-    public int hash;
+    public long hash;
 
     public TLObject deserializeResponse(AbstractSerializedData abstractSerializedData, int i, boolean z) {
         return TLRPC$account_Themes.TLdeserialize(abstractSerializedData, i, z);
@@ -12,6 +12,6 @@ public class TLRPC$TL_account_getThemes extends TLObject {
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
         abstractSerializedData.writeString(this.format);
-        abstractSerializedData.writeInt32(this.hash);
+        abstractSerializedData.writeInt64(this.hash);
     }
 }
