+++ sources/org/telegram/tgnet/TLRPC$TL_channelFull_layer103.java
@@ -13,7 +13,7 @@ public class TLRPC$TL_channelFull_layer103 extends TLRPC$TL_channelFull {
         this.hidden_prehistory = (readInt32 & 1024) != 0;
         this.can_view_stats = (readInt32 & 4096) != 0;
         this.can_set_location = (readInt32 & 65536) != 0;
-        this.id = abstractSerializedData.readInt32(z);
+        this.id = (long) abstractSerializedData.readInt32(z);
         this.about = abstractSerializedData.readString(z);
         if ((this.flags & 1) != 0) {
             this.participants_count = abstractSerializedData.readInt32(z);
@@ -52,7 +52,7 @@ public class TLRPC$TL_channelFull_layer103 extends TLRPC$TL_channelFull {
                 }
             }
             if ((this.flags & 16) != 0) {
-                this.migrated_from_chat_id = abstractSerializedData.readInt32(z);
+                this.migrated_from_chat_id = (long) abstractSerializedData.readInt32(z);
             }
             if ((this.flags & 16) != 0) {
                 this.migrated_from_max_id = abstractSerializedData.readInt32(z);
@@ -70,7 +70,7 @@ public class TLRPC$TL_channelFull_layer103 extends TLRPC$TL_channelFull {
                 this.folder_id = abstractSerializedData.readInt32(z);
             }
             if ((this.flags & 16384) != 0) {
-                this.linked_chat_id = abstractSerializedData.readInt32(z);
+                this.linked_chat_id = (long) abstractSerializedData.readInt32(z);
             }
             if ((this.flags & 32768) != 0) {
                 this.location = TLRPC$ChannelLocation.TLdeserialize(abstractSerializedData, abstractSerializedData.readInt32(z), z);
@@ -96,7 +96,7 @@ public class TLRPC$TL_channelFull_layer103 extends TLRPC$TL_channelFull {
         int i6 = this.can_set_location ? i5 | 65536 : i5 & -65537;
         this.flags = i6;
         abstractSerializedData.writeInt32(i6);
-        abstractSerializedData.writeInt32(this.id);
+        abstractSerializedData.writeInt32((int) this.id);
         abstractSerializedData.writeString(this.about);
         if ((this.flags & 1) != 0) {
             abstractSerializedData.writeInt32(this.participants_count);
@@ -126,7 +126,7 @@ public class TLRPC$TL_channelFull_layer103 extends TLRPC$TL_channelFull {
             this.bot_info.get(i7).serializeToStream(abstractSerializedData);
         }
         if ((this.flags & 16) != 0) {
-            abstractSerializedData.writeInt32(this.migrated_from_chat_id);
+            abstractSerializedData.writeInt32((int) this.migrated_from_chat_id);
         }
         if ((this.flags & 16) != 0) {
             abstractSerializedData.writeInt32(this.migrated_from_max_id);
@@ -144,7 +144,7 @@ public class TLRPC$TL_channelFull_layer103 extends TLRPC$TL_channelFull {
             abstractSerializedData.writeInt32(this.folder_id);
         }
         if ((this.flags & 16384) != 0) {
-            abstractSerializedData.writeInt32(this.linked_chat_id);
+            abstractSerializedData.writeInt32((int) this.linked_chat_id);
         }
         if ((this.flags & 32768) != 0) {
             this.location.serializeToStream(abstractSerializedData);
