+++ sources/org/telegram/tgnet/TLRPC$TL_channelParticipantAdmin_layer103.java
@@ -14,11 +14,11 @@ public class TLRPC$TL_channelParticipantAdmin_layer103 extends TLRPC$TL_channelP
         this.self = z2;
         TLRPC$TL_peerUser tLRPC$TL_peerUser = new TLRPC$TL_peerUser();
         this.peer = tLRPC$TL_peerUser;
-        tLRPC$TL_peerUser.user_id = abstractSerializedData.readInt32(z);
+        tLRPC$TL_peerUser.user_id = (long) abstractSerializedData.readInt32(z);
         if ((this.flags & 2) != 0) {
-            this.inviter_id = abstractSerializedData.readInt32(z);
+            this.inviter_id = (long) abstractSerializedData.readInt32(z);
         }
-        this.promoted_by = abstractSerializedData.readInt32(z);
+        this.promoted_by = (long) abstractSerializedData.readInt32(z);
         this.date = abstractSerializedData.readInt32(z);
         this.admin_rights = TLRPC$TL_chatAdminRights.TLdeserialize(abstractSerializedData, abstractSerializedData.readInt32(z), z);
     }
@@ -30,11 +30,11 @@ public class TLRPC$TL_channelParticipantAdmin_layer103 extends TLRPC$TL_channelP
         int i2 = this.self ? i | 2 : i & -3;
         this.flags = i2;
         abstractSerializedData.writeInt32(i2);
-        abstractSerializedData.writeInt32(this.peer.user_id);
+        abstractSerializedData.writeInt32((int) this.peer.user_id);
         if ((this.flags & 2) != 0) {
-            abstractSerializedData.writeInt32(this.inviter_id);
+            abstractSerializedData.writeInt32((int) this.inviter_id);
         }
-        abstractSerializedData.writeInt32(this.promoted_by);
+        abstractSerializedData.writeInt32((int) this.promoted_by);
         abstractSerializedData.writeInt32(this.date);
         this.admin_rights.serializeToStream(abstractSerializedData);
     }
