+++ sources/org/telegram/tgnet/TLRPC$TL_channelParticipantAdmin_layer92.java
@@ -13,9 +13,9 @@ public class TLRPC$TL_channelParticipantAdmin_layer92 extends TLRPC$TL_channelPa
         this.can_edit = z2;
         TLRPC$TL_peerUser tLRPC$TL_peerUser = new TLRPC$TL_peerUser();
         this.peer = tLRPC$TL_peerUser;
-        tLRPC$TL_peerUser.user_id = abstractSerializedData.readInt32(z);
-        this.inviter_id = abstractSerializedData.readInt32(z);
-        this.promoted_by = abstractSerializedData.readInt32(z);
+        tLRPC$TL_peerUser.user_id = (long) abstractSerializedData.readInt32(z);
+        this.inviter_id = (long) abstractSerializedData.readInt32(z);
+        this.promoted_by = (long) abstractSerializedData.readInt32(z);
         this.date = abstractSerializedData.readInt32(z);
         TLRPC$TL_channelAdminRights_layer92 TLdeserialize = TLRPC$TL_channelAdminRights_layer92.TLdeserialize(abstractSerializedData, abstractSerializedData.readInt32(z), z);
         this.admin_rights_layer92 = TLdeserialize;
@@ -27,9 +27,9 @@ public class TLRPC$TL_channelParticipantAdmin_layer92 extends TLRPC$TL_channelPa
         int i = this.can_edit ? this.flags | 1 : this.flags & -2;
         this.flags = i;
         abstractSerializedData.writeInt32(i);
-        abstractSerializedData.writeInt32(this.peer.user_id);
-        abstractSerializedData.writeInt32(this.inviter_id);
-        abstractSerializedData.writeInt32(this.promoted_by);
+        abstractSerializedData.writeInt32((int) this.peer.user_id);
+        abstractSerializedData.writeInt32((int) this.inviter_id);
+        abstractSerializedData.writeInt32((int) this.promoted_by);
         abstractSerializedData.writeInt32(this.date);
         this.admin_rights_layer92.serializeToStream(abstractSerializedData);
     }
