+++ sources/org/telegram/tgnet/TLRPC$TL_channels_getParticipants.java
@@ -4,7 +4,7 @@ public class TLRPC$TL_channels_getParticipants extends TLObject {
     public static int constructor = NUM;
     public TLRPC$InputChannel channel;
     public TLRPC$ChannelParticipantsFilter filter;
-    public int hash;
+    public long hash;
     public int limit;
     public int offset;
 
@@ -18,6 +18,6 @@ public class TLRPC$TL_channels_getParticipants extends TLObject {
         this.filter.serializeToStream(abstractSerializedData);
         abstractSerializedData.writeInt32(this.offset);
         abstractSerializedData.writeInt32(this.limit);
-        abstractSerializedData.writeInt32(this.hash);
+        abstractSerializedData.writeInt64(this.hash);
     }
 }
