+++ sources/org/telegram/tgnet/TLRPC$TL_chatAdminWithInvites.java
@@ -2,7 +2,7 @@ package org.telegram.tgnet;
 
 public class TLRPC$TL_chatAdminWithInvites extends TLObject {
     public static int constructor = -NUM;
-    public int admin_id;
+    public long admin_id;
     public int invites_count;
     public int revoked_invites_count;
 
@@ -19,14 +19,14 @@ public class TLRPC$TL_chatAdminWithInvites extends TLObject {
     }
 
     public void readParams(AbstractSerializedData abstractSerializedData, boolean z) {
-        this.admin_id = abstractSerializedData.readInt32(z);
+        this.admin_id = abstractSerializedData.readInt64(z);
         this.invites_count = abstractSerializedData.readInt32(z);
         this.revoked_invites_count = abstractSerializedData.readInt32(z);
     }
 
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
-        abstractSerializedData.writeInt32(this.admin_id);
+        abstractSerializedData.writeInt64(this.admin_id);
         abstractSerializedData.writeInt32(this.invites_count);
         abstractSerializedData.writeInt32(this.revoked_invites_count);
     }
