+++ sources/org/telegram/tgnet/TLRPC$TL_chatForbidden.java
@@ -4,13 +4,13 @@ public class TLRPC$TL_chatForbidden extends TLRPC$Chat {
     public static int constructor = NUM;
 
     public void readParams(AbstractSerializedData abstractSerializedData, boolean z) {
-        this.id = abstractSerializedData.readInt32(z);
+        this.id = abstractSerializedData.readInt64(z);
         this.title = abstractSerializedData.readString(z);
     }
 
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
-        abstractSerializedData.writeInt32(this.id);
+        abstractSerializedData.writeInt64(this.id);
         abstractSerializedData.writeString(this.title);
     }
 }
