+++ sources/org/telegram/tgnet/TLRPC$TL_chatFull.java
@@ -9,7 +9,7 @@ public class TLRPC$TL_chatFull extends TLRPC$ChatFull {
         int i = 0;
         this.can_set_username = (readInt32 & 128) != 0;
         this.has_scheduled = (readInt32 & 256) != 0;
-        this.id = abstractSerializedData.readInt32(z);
+        this.id = abstractSerializedData.readInt64(z);
         this.about = abstractSerializedData.readString(z);
         this.participants = TLRPC$ChatParticipants.TLdeserialize(abstractSerializedData, abstractSerializedData.readInt32(z), z);
         if ((this.flags & 4) != 0) {
@@ -65,7 +65,7 @@ public class TLRPC$TL_chatFull extends TLRPC$ChatFull {
         int i2 = this.has_scheduled ? i | 256 : i & -257;
         this.flags = i2;
         abstractSerializedData.writeInt32(i2);
-        abstractSerializedData.writeInt32(this.id);
+        abstractSerializedData.writeInt64(this.id);
         abstractSerializedData.writeString(this.about);
         this.participants.serializeToStream(abstractSerializedData);
         if ((this.flags & 4) != 0) {
