+++ sources/org/telegram/tgnet/TLRPC$TL_chatParticipants.java
@@ -4,7 +4,7 @@ public class TLRPC$TL_chatParticipants extends TLRPC$ChatParticipants {
     public static int constructor = NUM;
 
     public void readParams(AbstractSerializedData abstractSerializedData, boolean z) {
-        this.chat_id = abstractSerializedData.readInt32(z);
+        this.chat_id = abstractSerializedData.readInt64(z);
         int readInt32 = abstractSerializedData.readInt32(z);
         int i = 0;
         if (readInt32 == NUM) {
@@ -26,7 +26,7 @@ public class TLRPC$TL_chatParticipants extends TLRPC$ChatParticipants {
 
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
-        abstractSerializedData.writeInt32(this.chat_id);
+        abstractSerializedData.writeInt64(this.chat_id);
         abstractSerializedData.writeInt32(NUM);
         int size = this.participants.size();
         abstractSerializedData.writeInt32(size);
