+++ sources/org/telegram/tgnet/TLRPC$TL_chatParticipantsForbidden_old.java
@@ -4,11 +4,11 @@ public class TLRPC$TL_chatParticipantsForbidden_old extends TLRPC$TL_chatPartici
     public static int constructor = NUM;
 
     public void readParams(AbstractSerializedData abstractSerializedData, boolean z) {
-        this.chat_id = abstractSerializedData.readInt32(z);
+        this.chat_id = (long) abstractSerializedData.readInt32(z);
     }
 
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
-        abstractSerializedData.writeInt32(this.chat_id);
+        abstractSerializedData.writeInt32((int) this.chat_id);
     }
 }
