+++ sources/org/telegram/tgnet/TLRPC$TL_documentEncrypted_old.java
@@ -6,7 +6,7 @@ public class TLRPC$TL_documentEncrypted_old extends TLRPC$TL_document {
     public void readParams(AbstractSerializedData abstractSerializedData, boolean z) {
         this.id = abstractSerializedData.readInt64(z);
         this.access_hash = abstractSerializedData.readInt64(z);
-        this.user_id = abstractSerializedData.readInt32(z);
+        this.user_id = (long) abstractSerializedData.readInt32(z);
         this.date = abstractSerializedData.readInt32(z);
         this.file_name = abstractSerializedData.readString(z);
         this.mime_type = abstractSerializedData.readString(z);
@@ -21,7 +21,7 @@ public class TLRPC$TL_documentEncrypted_old extends TLRPC$TL_document {
         abstractSerializedData.writeInt32(constructor);
         abstractSerializedData.writeInt64(this.id);
         abstractSerializedData.writeInt64(this.access_hash);
-        abstractSerializedData.writeInt32(this.user_id);
+        abstractSerializedData.writeInt32((int) this.user_id);
         abstractSerializedData.writeInt32(this.date);
         abstractSerializedData.writeString(this.file_name);
         abstractSerializedData.writeString(this.mime_type);
