+++ sources/org/telegram/tgnet/TLRPC$TL_encryptedChat_old.java
@@ -7,8 +7,8 @@ public class TLRPC$TL_encryptedChat_old extends TLRPC$TL_encryptedChat {
         this.id = abstractSerializedData.readInt32(z);
         this.access_hash = abstractSerializedData.readInt64(z);
         this.date = abstractSerializedData.readInt32(z);
-        this.admin_id = abstractSerializedData.readInt32(z);
-        this.participant_id = abstractSerializedData.readInt32(z);
+        this.admin_id = (long) abstractSerializedData.readInt32(z);
+        this.participant_id = (long) abstractSerializedData.readInt32(z);
         this.g_a_or_b = abstractSerializedData.readByteArray(z);
         this.nonce = abstractSerializedData.readByteArray(z);
         this.key_fingerprint = abstractSerializedData.readInt64(z);
@@ -19,8 +19,8 @@ public class TLRPC$TL_encryptedChat_old extends TLRPC$TL_encryptedChat {
         abstractSerializedData.writeInt32(this.id);
         abstractSerializedData.writeInt64(this.access_hash);
         abstractSerializedData.writeInt32(this.date);
-        abstractSerializedData.writeInt32(this.admin_id);
-        abstractSerializedData.writeInt32(this.participant_id);
+        abstractSerializedData.writeInt32((int) this.admin_id);
+        abstractSerializedData.writeInt32((int) this.participant_id);
         abstractSerializedData.writeByteArray(this.g_a_or_b);
         abstractSerializedData.writeByteArray(this.nonce);
         abstractSerializedData.writeInt64(this.key_fingerprint);
