+++ sources/org/telegram/tgnet/TLRPC$TL_importedContact.java
@@ -3,7 +3,7 @@ package org.telegram.tgnet;
 public class TLRPC$TL_importedContact extends TLObject {
     public static int constructor = -NUM;
     public long client_id;
-    public int user_id;
+    public long user_id;
 
     public static TLRPC$TL_importedContact TLdeserialize(AbstractSerializedData abstractSerializedData, int i, boolean z) {
         if (constructor == i) {
@@ -18,13 +18,13 @@ public class TLRPC$TL_importedContact extends TLObject {
     }
 
     public void readParams(AbstractSerializedData abstractSerializedData, boolean z) {
-        this.user_id = abstractSerializedData.readInt32(z);
+        this.user_id = abstractSerializedData.readInt64(z);
         this.client_id = abstractSerializedData.readInt64(z);
     }
 
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
-        abstractSerializedData.writeInt32(this.user_id);
+        abstractSerializedData.writeInt64(this.user_id);
         abstractSerializedData.writeInt64(this.client_id);
     }
 }
